feat: opcode cycle counts
This commit is AI slop.

I don't like this implementation, but it's basically the best way to do it without major refactors to the CPU emulator, so we're just gonna live with it for now. I'd eventually like to make a biblically-accurate CPU emulator, but right now I just want to get the PPU working well enough to see the emulator work on Super Mario Bros. (OG), which still is unplayable.